[docs] 경력사항 DB 연동 완료
