+++ package.json
@@ -18,7 +18,7 @@
     "@fullcalendar/list": "^6.1.19",
     "@fullcalendar/react": "^6.1.19",
     "@fullcalendar/timegrid": "^6.1.19",
-    "@supabase/supabase-js": "^2.57.4",
+    "@supabase/supabase-js": "^2.75.0",
     "@uiw/react-markdown-preview": "^5.1.5",
     "@uiw/react-md-editor": "^4.0.8",
     "antd": "^5.27.4",
